+++ src/views/Authorization/Role/components/Write.vue
@@ -29,11 +29,6 @@ const formSchema = ref<FormSchema[]>([
     component: 'Input'
   },
   {
-    field: 'role',
-    label: t('role.role'),
-    component: 'Input'
-  },
-  {
     field: 'status',
     label: t('menu.status'),
     component: 'Select',
@@ -68,6 +63,7 @@ const formSchema = ref<FormSchema[]>([
                     show-checkbox
                     node-key="id"
                     highlight-current
+                    check-strictly
                     expand-on-click-node={false}
                     data={treeData.value}
                     onNode-click={nodeClick}
@@ -80,10 +76,10 @@ const formSchema = ref<FormSchema[]>([
                   </ElTree>
                 </div>
                 <div class="flex-1">
-                  {unref(currentTreeData) && unref(currentTreeData)?.permission ? (
+                  {unref(currentTreeData) && unref(currentTreeData)?.permissionList ? (
                     <ElCheckboxGroup v-model={unref(currentTreeData).meta.permission}>
-                      {unref(currentTreeData)?.permission.map((v: string) => {
-                        return <ElCheckbox label={v} />
+                      {unref(currentTreeData)?.permissionList.map((v: any) => {
+                        return <ElCheckbox label={v.value}>{v.label}</ElCheckbox>
                       })}
                     </ElCheckboxGroup>
                   ) : null}
@@ -153,10 +149,7 @@ const submit = async () => {
   })
   if (valid) {
     const formData = await getFormData()
-    const checkedKeys = [
-      ...(unref(treeRef)?.getCheckedKeys() || []),
-      ...(unref(treeRef)?.getHalfCheckedKeys() || [])
-    ]
+    const checkedKeys = unref(treeRef)?.getCheckedKeys() || []
     const data = filter(unref(treeData), (item: any) => {
       return checkedKeys.includes(item.id)
     })